+++ webapp/src/commonComponents/grids/lostDogsGrid.test.tsx
@@ -4,19 +4,12 @@
 
 import React from "react";
 import { render, unmountComponentAtNode } from "react-dom";
-import { act, createRenderer } from "react-dom/test-utils";
+import { act } from "react-dom/test-utils";
 import { isNull } from "lodash";
-import { Button } from "@material-ui/core";
-import ShelterDogGrid from "./shelterDogGrid";
-import { otherDog, sheltertestDog, testDogList } from "../../dog/dogTesting";
+import LostDogsGrid from "./lostDogsGrid";
+import { testDogList, testLostDogList } from "../../dog/dogTesting";
 import { Provider } from "react-redux";
-import {
-  Redirect,
-  Route,
-  BrowserRouter as Router,
-  Switch,
-  useHistory,
-} from "react-router-dom";
+import { Route, BrowserRouter as Router } from "react-router-dom";
 import { store } from "../../app/store";
 
 let container: HTMLDivElement | null = null;
@@ -41,7 +34,7 @@ it("Rendered list of dog cards", () => {
       <Provider store={store}>
         <Router>
           <Route path="/">
-            <ShelterDogGrid dogs={testDogList} id={-1} cookies={{}} path={""} />
+            <LostDogsGrid dogs={testLostDogList} id={-1} cookies={{}} path={""} />
           </Route>
         </Router>
       </Provider>,
@@ -52,7 +45,7 @@ it("Rendered list of dog cards", () => {
   //container?.getElementsByTagName("registerButton")
   expect(
     container?.getElementsByClassName("tile").length
-  ).toBeGreaterThanOrEqual(testDogList.length);
+  ).toBeGreaterThanOrEqual(testLostDogList.length);
   //check if dog named alex was rendered
-  expect(container?.getElementsByClassName("Alexa").length).toEqual(2);
+  expect(container?.getElementsByClassName("Alex").length).toEqual(1);
 });
