+++ webapp/src/commonComponents/grids/lostDogsGrid.tsx
@@ -3,18 +3,26 @@ import GridList from "@material-ui/core/GridList";
 import GridListTile from "@material-ui/core/GridListTile";
 import GridListTileBar from "@material-ui/core/GridListTileBar";
 import IconButton from "@material-ui/core/IconButton";
-import { IShelterDog, IShelterDogWithPicture } from "../../dog/dogInterfaces";
-import {
-  CheckBoxOutlineBlank,
-  CheckBoxTwoTone,
-  Delete,
-} from "@material-ui/icons";
+import InfoIcon from "@material-ui/icons/Info";
+import { ILostDogWithPicture } from "../../dog/dogInterfaces";
+import { Edit } from "@material-ui/icons";
+import { store } from "../../app/store";
+import { fetchOneDogThunk } from "../../app/actions";
 import { useCookies } from "react-cookie";
-import { Checkbox } from "@material-ui/core";
 
-export default function ShelterDogGrid(props: any) {
-  const dogs = props.dogs as IShelterDogWithPicture[]; // eslint-disable-next-line
+export default function LostDogsGrid(props: any) {
+  const dogs = props.dogs as ILostDogWithPicture[]; // eslint-disable-next-line
   const [cookies, setCookie, removeCookie] = useCookies();
+  const redirectToDogDetailsOrEdit = (id: number) => {
+    store.dispatch(
+      fetchOneDogThunk({
+        id: id as number,
+        cookies: cookies,
+      })
+    );
+    props.redirectToDogDetailsOrEdit(id);
+  };
+
   return (
     <GridList
       cols={3}
@@ -22,14 +30,15 @@ export default function ShelterDogGrid(props: any) {
       style={{ margin: "0", width: "100%", display: "flex" }}
     >
       {dogs.map(
-        (dog: IShelterDogWithPicture) =>
+        (dog: ILostDogWithPicture) =>
           dog.picture && (
             <GridListTile
               key={dog.id}
-              style={{ height: "300px" }}
+              style={{ minHeight: "300px", display: "stretch" }}
               className="tile"
             >
               <img
+                style={{ minHeight: "300px", display: "stretch" }}
                 src={`data:${dog.picture.fileType};base64,${
                   dog.picture.data as ArrayBuffer
                 }`}
@@ -38,13 +47,20 @@ export default function ShelterDogGrid(props: any) {
               <GridListTileBar
                 className={dog.name}
                 title={dog.name}
-                subtitle={<span>{dog.breed}</span>}
+                subtitle={
+                  <span>
+                    {dog.isFound ? "Found" : "Lost in " + dog.location.city}
+                  </span>
+                }
                 actionIcon={
                   <IconButton
                     aria-label={`info about ${dog.name}`}
                     style={{ color: "rgba(255, 255, 255, 0.54)" }}
+                    onClick={() => {
+                      redirectToDogDetailsOrEdit(dog.id as number);
+                    }}
                   >
-                    <Checkbox />
+                    {props.path === "/dogs" ? <InfoIcon /> : <Edit />}
                   </IconButton>
                 }
               />
@@ -53,7 +69,7 @@ export default function ShelterDogGrid(props: any) {
       )}
       {dogs.length <= 2 && dogs.length > 0 && dogs[0] && dogs[0].picture && (
         <GridListTile
-          key={"dog.id1"}
+          key={"dogid1"}
           style={{ height: "300px" }}
           className="tile"
         >
@@ -67,7 +83,7 @@ export default function ShelterDogGrid(props: any) {
       )}
       {dogs.length == 1 && dogs[0] && dogs[0].picture && (
         <GridListTile
-          key={"dog.id2"}
+          key={"dogid2"}
           style={{ height: "300px" }}
           className="tile"
         >
