+++ webapp/src/commonComponents/grids/shelterDogsGrid.test.tsx
@@ -4,13 +4,20 @@
 
 import React from "react";
 import { render, unmountComponentAtNode } from "react-dom";
-import { act } from "react-dom/test-utils";
+import { act, createRenderer } from "react-dom/test-utils";
 import { isNull } from "lodash";
+import { Button } from "@material-ui/core";
+import ShelterDogsGrid from "./shelterDogsGrid";
+import { otherDog, sheltertestDog, testDogList } from "../../dog/dogTesting";
 import { Provider } from "react-redux";
-import { Route, BrowserRouter as Router } from "react-router-dom";
+import {
+  Redirect,
+  Route,
+  BrowserRouter as Router,
+  Switch,
+  useHistory,
+} from "react-router-dom";
 import { store } from "../../app/store";
-import SheltersGrid from "./sheltersGrid";
-import { exampleShelter, initShelter } from "../../shelter/shelterTesting";
 
 let container: HTMLDivElement | null = null;
 beforeEach(() => {
@@ -28,18 +35,13 @@ afterEach(() => {
   }
 });
 
-it("Rendered list of shelters", () => {
+it("Rendered list of dog cards", () => {
   act(() => {
     render(
       <Provider store={store}>
         <Router>
-          <Route path="/shelters">
-            <SheltersGrid
-              shelters={[initShelter, exampleShelter]}
-              id={-1}
-              cookies={{}}
-              path={""}
-            />
+          <Route path="/">
+            <ShelterDogsGrid dogs={testDogList} id={-1} cookies={{}} path={""} />
           </Route>
         </Router>
       </Provider>,
@@ -50,9 +52,7 @@ it("Rendered list of shelters", () => {
   //container?.getElementsByTagName("registerButton")
   expect(
     container?.getElementsByClassName("tile").length
-  ).toBeGreaterThanOrEqual(0);
+  ).toBeGreaterThanOrEqual(testDogList.length);
   //check if dog named alex was rendered
-  expect(
-    container?.getElementsByClassName("The Best Shelter").length
-  ).toBeGreaterThanOrEqual(0);
+  expect(container?.getElementsByClassName("Alexa").length).toEqual(2);
 });
