+++ webapp/src/commonComponents/grids/shelterDogsGrid.tsx
@@ -3,26 +3,13 @@ import GridList from "@material-ui/core/GridList";
 import GridListTile from "@material-ui/core/GridListTile";
 import GridListTileBar from "@material-ui/core/GridListTileBar";
 import IconButton from "@material-ui/core/IconButton";
-import InfoIcon from "@material-ui/icons/Info";
-import { ILostDogWithPicture } from "../../dog/dogInterfaces";
-import { Edit } from "@material-ui/icons";
-import { store } from "../../app/store";
-import { fetchOneDogThunk } from "../../app/actions";
+import { IShelterDogWithPicture } from "../../dog/dogInterfaces";
 import { useCookies } from "react-cookie";
+import { Checkbox } from "@material-ui/core";
 
-export default function ImageGrid(props: any) {
-  const dogs = props.dogs as ILostDogWithPicture[]; // eslint-disable-next-line
+export default function ShelterDogsGrid(props: any) {
+  const dogs = props.dogs as IShelterDogWithPicture[]; // eslint-disable-next-line
   const [cookies, setCookie, removeCookie] = useCookies();
-  const redirectToDogDetailsOrEdit = (id: number) => {
-    store.dispatch(
-      fetchOneDogThunk({
-        id: id as number,
-        cookies: cookies,
-      })
-    );
-    props.redirectToDogDetailsOrEdit(id);
-  };
-
   return (
     <GridList
       cols={3}
@@ -30,15 +17,14 @@ export default function ImageGrid(props: any) {
       style={{ margin: "0", width: "100%", display: "flex" }}
     >
       {dogs.map(
-        (dog: ILostDogWithPicture) =>
+        (dog: IShelterDogWithPicture) =>
           dog.picture && (
             <GridListTile
               key={dog.id}
-              style={{ minHeight: "300px", display: "stretch" }}
+              style={{ height: "300px" }}
               className="tile"
             >
               <img
-                style={{ minHeight: "300px", display: "stretch" }}
                 src={`data:${dog.picture.fileType};base64,${
                   dog.picture.data as ArrayBuffer
                 }`}
@@ -47,20 +33,13 @@ export default function ImageGrid(props: any) {
               <GridListTileBar
                 className={dog.name}
                 title={dog.name}
-                subtitle={
-                  <span>
-                    {dog.isFound ? "Found" : "Lost in " + dog.location.city}
-                  </span>
-                }
+                subtitle={<span>{dog.breed}</span>}
                 actionIcon={
                   <IconButton
                     aria-label={`info about ${dog.name}`}
                     style={{ color: "rgba(255, 255, 255, 0.54)" }}
-                    onClick={() => {
-                      redirectToDogDetailsOrEdit(dog.id as number);
-                    }}
                   >
-                    {props.path === "/dogs" ? <InfoIcon /> : <Edit />}
+                    <Checkbox />
                   </IconButton>
                 }
               />
@@ -69,7 +48,7 @@ export default function ImageGrid(props: any) {
       )}
       {dogs.length <= 2 && dogs.length > 0 && dogs[0] && dogs[0].picture && (
         <GridListTile
-          key={"dogid1"}
+          key={"dog.id1"}
           style={{ height: "300px" }}
           className="tile"
         >
@@ -83,7 +62,7 @@ export default function ImageGrid(props: any) {
       )}
       {dogs.length == 1 && dogs[0] && dogs[0].picture && (
         <GridListTile
-          key={"dogid2"}
+          key={"dog.id2"}
           style={{ height: "300px" }}
           className="tile"
         >
