+++ webapp/src/commonComponents/grids/sheltersInfoGrid.test.tsx
@@ -6,11 +6,11 @@ import React from "react";
 import { render, unmountComponentAtNode } from "react-dom";
 import { act } from "react-dom/test-utils";
 import { isNull } from "lodash";
-import ImageGrid from "./imageGrid";
-import { testDogList, testLostDogList } from "../../dog/dogTesting";
 import { Provider } from "react-redux";
 import { Route, BrowserRouter as Router } from "react-router-dom";
 import { store } from "../../app/store";
+import SheltersInfoGrid from "./sheltersInfoGrid";
+import { exampleShelter, initShelter } from "../../shelter/shelterTesting";
 
 let container: HTMLDivElement | null = null;
 beforeEach(() => {
@@ -28,13 +28,18 @@ afterEach(() => {
   }
 });
 
-it("Rendered list of dog cards", () => {
+it("Rendered list of shelters", () => {
   act(() => {
     render(
       <Provider store={store}>
         <Router>
-          <Route path="/">
-            <ImageGrid dogs={testLostDogList} id={-1} cookies={{}} path={""} />
+          <Route path="/shelters">
+            <SheltersInfoGrid
+              shelters={[initShelter, exampleShelter]}
+              id={-1}
+              cookies={{}}
+              path={""}
+            />
           </Route>
         </Router>
       </Provider>,
@@ -45,7 +50,9 @@ it("Rendered list of dog cards", () => {
   //container?.getElementsByTagName("registerButton")
   expect(
     container?.getElementsByClassName("tile").length
-  ).toBeGreaterThanOrEqual(testLostDogList.length);
+  ).toBeGreaterThanOrEqual(0);
   //check if dog named alex was rendered
-  expect(container?.getElementsByClassName("Alex").length).toEqual(1);
+  expect(
+    container?.getElementsByClassName("The Best Shelter").length
+  ).toBeGreaterThanOrEqual(0);
 });
