+++ webapp/src/registerLogin/Login.tsx
@@ -109,7 +109,7 @@ export default function Login() {
       setCookie(config.cookies.userType, loginInfo?.userType, { path: "/" });
       setCookie(config.cookies.userId, loginInfo?.id, { path: "/" });
       store.dispatch(clearLoginInformation());
-      if (loginInfo?.userType == "Shelter") {
+      if (loginInfo?.userType === "Shelter") {
         history.push("/shelterdogs");
       } else {
         history.push("/dogs");
@@ -120,7 +120,7 @@ export default function Login() {
   //THIS makes the web app skip the log in
   useEffect(() => {
     if (cookies[config.cookies.userType] !== undefined) {
-      if (cookies[config.cookies.userType] == "Shelter") {
+      if (cookies[config.cookies.userType] === "Shelter") {
         history.push("/shelterdogs");
       } else {
         history.push("/dogs");
