+++ webapp/src/shelter/registerShelterDog/registerShelterDog.tsx
@@ -35,7 +35,7 @@ import { store } from "../../app/store";
 import { useCookies } from "react-cookie";
 import { useHistory } from "react-router-dom";
 import config from "../../config/config";
-import ImageUpload from "../../commonComponents/ImageUpload";
+import ImageUpload from "../../commonComponents/imageUploadForm";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
